djangoFunction/requirements.txt
azure-functions
django
whitenoise
psycopg2-binary
